# Conflicts:
#	src/main/java/com/itstep/flightticketsbooking/api/controller/AdminRestController.java
#	src/main/java/com/itstep/flightticketsbooking/api/dto/FlightDto.java
#	src/main/resources/templates/adminPage.html

+++ src/main/java/com/itstep/flightticketsbooking/repository/PassengerRepository.java
@@ -3,6 +3,6 @@ package com.itstep.flightticketsbooking.repository;
 import com.itstep.flightticketsbooking.entity.Passenger;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PassengerRepository extends JpaRepository<Passenger, Integer> {
+public interface PassengerRepository extends JpaRepository<Passenger, Long> {
 
 }
